safari-534.53-branch/LayoutTests/http/tests/cookies/double-quoted-value-with-semi-colon-expected.txt

Test for <rdar://problem/5666078> Cookie parsing terminates at the first semicolon, ignoring quotes (16699)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Check that setting a cookie with a semi-colon in a duoble-quoted value works
PASS cookie is 'disorder="477beccb;richard"'.
PASS successfullyParsed is true

TEST COMPLETE
